Refactors test for data gap hendeling to rely on _pwv_date function
